pyproject.toml
[project]
name = "website-intelligence-platform"
version = "1.0.0"
description = "AI-powered website navigation intelligence extraction and memory management platform"
authors = [
    {name = "Your Company", email = "support@yourcompany.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9,<3.12"
keywords = ["web-crawler", "ai", "navigation", "memory-management", "b2b", "multi-tenant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",

    # Database
    "motor>=3.3.2",
    "pymongo>=4.6.0",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",

    # Web Crawling
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "aiohttp>=3.9.1",
    "playwright>=1.40.0",
    "selenium>=4.15.2",
    "requests>=2.31.0",
    "httpx>=0.25.2",

    # AI/ML
    "openai>=1.3.5",
    "anthropic>=0.7.0",
    "langchain>=0.0.340",
    "tiktoken>=0.5.1",
    "sentence-transformers>=2.2.2",

    # Vector Stores
    "chromadb>=0.4.18",
    "pinecone-client>=2.2.4",
    "qdrant-client>=1.7.0",

    # Task Queue
    "celery>=5.3.4",
    "flower>=2.0.1",

    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",

    # Development
    "ipython>=8.17.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
    "pre-commit>=3.5.0",
    "ipdb>=0.13.13",
    "rich>=13.7.0",
]

test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "faker>=20.1.0",
    "factory-boy>=3.3.0",
    "httpx>=0.25.2",
]

docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.14",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/yourcompany/website-intelligence-platform"
Documentation = "https://docs.yourcompany.com/wip"
Repository = "https://github.com/yourcompany/website-intelligence-platform"
Issues = "https://github.com/yourcompany/website-intelligence-platform/issues"

[project.scripts]
wip = "src.cli:main"
wip-api = "src.api.main:run"
wip-crawler = "src.crawlers.cli:main"
wip-scheduler = "src.services.scheduler:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
preview = true

[tool.uv.pip]
system = false
break-system-packages = false

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  | migrations
  | .venv
  | venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip() without an explicit strict= parameter
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    "migrations",
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_generics = false
ignore_missing_imports = true
exclude = [
    "migrations/",
    "venv/",
    ".venv/",
    "build/",
    "dist/",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--asyncio-mode=auto",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "class .*\\(Protocol\\):",
]